requirements.txt
streamlit==1.29.0
pandas==2.1.0
plotly==5.18.0
wordcloud==1.9.3
streamlit-aggrid==0.3.4.post3
biopython==1.81
googletrans==4.0.0-rc1
keybert==0.8.0
sentence-transformers==2.2.2
torch>=2.0.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
python-dotenv==1.0.0
supabase==0.10.0